+++ src/layouts/Chat.vue
@@ -48,6 +48,7 @@ export default defineComponent({
     width: 250px;
     height: 100%;
     overflow-y: scroll;
+    border-right: 1px solid var(--el-border-color);
   }
 
   .chat {