+++ src/layouts/Hailuo.vue
@@ -1,6 +1,7 @@
 <template>
   <div class="main">
     <div class="config">
+      <video-navigator />
       <slot name="config" />
     </div>
     <div class="result">
@@ -19,13 +20,15 @@
 import { defineComponent } from 'vue';
 import { ElDrawer, ElButton } from 'element-plus';
 import { FontAwesomeIcon } from '@fortawesome/vue-fontawesome';
+import VideoNavigator from '@/components/common/VideoNavigator.vue';
 
 export default defineComponent({
   name: 'LayoutHailuo',
   components: {
     ElDrawer,
     ElButton,
-    FontAwesomeIcon
+    FontAwesomeIcon,
+    VideoNavigator
   },
   data() {
     return {