+++ src/layouts/Midjourney.vue
@@ -1,6 +1,7 @@
 <template>
   <div class="main">
     <div class="presets">
+      <image-navigator />
       <slot name="presets" />
     </div>
     <div class="operation">
@@ -19,13 +20,15 @@
 import { defineComponent } from 'vue';
 import { ElDrawer, ElButton } from 'element-plus';
 import { FontAwesomeIcon } from '@fortawesome/vue-fontawesome';
+import ImageNavigator from '@/components/common/ImageNavigator.vue';
 
 export default defineComponent({
   name: 'LayoutMidjourney',
   components: {
     ElDrawer,
     ElButton,
-    FontAwesomeIcon
+    FontAwesomeIcon,
+    ImageNavigator
   },
   data() {
     return {