+++ src/pages/console/application/List.vue
@@ -26,10 +26,10 @@
               </el-table-column>
               <el-table-column :label="$t('application.field.type')" width="80px">
                 <template #default="scope">
-                  <el-tag v-if="scope.row?.type === 'Period'" type="success">
+                  <el-tag v-if="scope.row?.type === 'Period'" type="success" effect="dark">
                     {{ $t('application.type.period') }}
                   </el-tag>
-                  <el-tag v-if="scope.row?.type === 'Usage'" type="primary">
+                  <el-tag v-if="scope.row?.type === 'Usage'" effect="dark">
                     {{ $t('application.type.usage') }}
                   </el-tag>
                 </template>
@@ -100,8 +100,8 @@ import { applicationOperator } from '@/operators';
 import Pagination from '@/components/common/Pagination.vue';
 import CopyToClipboard from '@/components/common/CopyToClipboard.vue';
 import { ElTable, ElRow, ElCol, ElTableColumn, ElCard, ElButton, ElTag } from 'element-plus';
-import { ROUTE_CONSOLE_APPLICATION_EXTRA, ROUTE_CONSOLE_APPLICATION_SUBSCRIBE } from '@/router/constants';
-import { IApplication, IApplicationListResponse, IApplicationType, ICredentialType, IService } from '@/models';
+import { ROUTE_CONSOLE_APPLICATION_SUBSCRIBE } from '@/router/constants';
+import { IApplication, IApplicationListResponse, ICredentialType, IService } from '@/models';
 
 interface IData {
   applications: IApplication[];
@@ -174,21 +174,12 @@ export default defineComponent({
   },
   methods: {
     onBuyMore(application: IApplication) {
-      if (application.type === IApplicationType.USAGE) {
-        this.$router.push({
-          name: ROUTE_CONSOLE_APPLICATION_EXTRA,
-          params: {
-            id: application.id
-          }
-        });
-      } else if (application.type === IApplicationType.PERIOD) {
-        this.$router.push({
-          name: ROUTE_CONSOLE_APPLICATION_SUBSCRIBE,
-          params: {
-            id: application.id
-          }
-        });
-      }
+      this.$router.push({
+        name: ROUTE_CONSOLE_APPLICATION_SUBSCRIBE,
+        params: {
+          id: application.id
+        }
+      });
     },
     onPageChange(page: number) {
       this.$router.push({