+++ src/store/common/models.ts
@@ -5,6 +5,7 @@ import { IChatdocState } from '../chatdoc/models';
 import { IQrartState } from '../qrart/models';
 import { ILumaState } from '../luma/models';
 import { IPikaState } from '../pika/models';
+import { IFluxState } from '../flux/models';
 import { IHailuoState } from '../hailuo/models';
 import { IHeadshotsState } from '../headshots/models';
 import { ISunoState } from '../suno/models';
@@ -36,6 +37,7 @@ export interface IRootState extends ICommonState {
   qrart: IQrartState;
   luma: ILumaState;
   pika: IPikaState;
+  flux: IFluxState;
   hailuo: IHailuoState;
   headshots: IHeadshotsState;
   suno: ISunoState;